Add lifestream-tools
Lifestream-tools is based on the idea of livecd-tools in Gentoo where it adds fixes for lifestream to work.
It takes it to the next level by pulling in all needed packagers as depends to make building a system for the users needs based on selecting the correct USE flag at build time.

Adds the lifestream user creation as wel, this is an idea in progress for adding user fixes such as stopping chrome for updating as policy.

Signed-off-by: Ian Jordan <immoloism@gmail.com>